Look at DialogFragment inflated layout before the dialog's
When using something like a BottomSheetDialog, because it reuses ids for
its child view, the dialog's view will be consider as the container
instead of the actual inflated container. To fix this, we should just
look at the inflated layout first.

RelNote: "You can now add fragment to `BottomSheetDialogFragment`s and
they will have a properly nested view hierarchy."
Test: DialogFragmentTest
Bug: 180021387

Change-Id: Ie62791677cb8771f22b19bdce5985426c669fba0