Align panes when folding feature is separating
Currently SlidingPaneLayout measures the side and content panes
correctly, but when the views are placed during onLayout() they don't
take into account the folding feature width (if any), so content could
be rendered under a folding feature if this has a >0px width.

This fix takes into account if there is a folding feature that
separates the panes and if so, uses the folding feature width as the
next view starting offset.

Test: tested on a dual-screen device and emulator with physical hinge
where its width is > 0px.
Tested as well in all foldable available emulators to check that no
wrong effects have been added.
Did run as well all unit and instrumentation tests already available.

Change-Id: I0e9f92936ae41f5d6a6ccf87da8fa02895e4d7fd