Remove support for specifying digest algorithm for certs
Android 12 introduced a new permission protection flag that allows
a set of trusted certs to be specified with the permission
declaration; if a requesting app is signed by one of these certs
the permission is granted to the app. Since the platform only
supports specifying the SHA-256 digest of the trusted certs this
commit requires that apps also use SHA-256 when declaring
their app-authenticator XML config file.

Bug: 171453012
Test: AppAuthenticatorTest
Change-Id: Iad202903ffa1267e2b7e6892d6a8beabfcf291b6