Improve flaky ImageAnalysis tests in CameraControllerFragmentTest
Require 2 analysis frame to verify ImageAnalysis is running or not. There is a possible race condition where posted frame can arrive after the analzyer is cleared. Not sure if this is the cause of the test failure but worth a try.

Also: removed the unecessary close() call. The image is always closed in the test app.

Bug: 172955130
Test: manual test and ./gradlew bOS
Change-Id: I31aab2cac1805c876b9484b37fa94470b9a73194