Add prunePackageData in AppSearchImpl.
Since we cannot access locked user's data and clear it when a package
got removed. When a user is unlocking, we should prune their data and
remove all data that doesn't belong to any of installed packages.

This will also help in handling the case where device gets
rebooted before the data is cleared for an uninstalled package.

Bug: 186151459
Test: AppSearchImplTest
Change-Id: Iba60685f394080708a95be709533339e3e91cf5c