Call setPageSize for LegacyPagingSource from .asPagingSourceFactory
There are currently some instances where we receive a PagingSource, but
do not check if it is an instance of LegacyPagingSource. This can cause
use-cases which explicitly call .asPagingSourceFactory.invoke() such as
Room to fail to set pageSize, which can cause an upgrade to paging3 to
inadvertently load a different number of items than expected.

Fixes: 181805612
Test: ./gradlew paging:paging-runtime:cC
Change-Id: Ic394fabf3e4712dcc9484eca3ddd2da2d8841a8b