Add a new CaptureVideo contract
The current TakeVideo contract attempts to return a thumbnail as a
Bitmap, but no intent action actually returns a Bitmap for a video so
the result always ends up being null, which is not useful. We should
instead return a boolean that indicates whether the video was
successfully recorded and saved.

RelNote: "The new `CaptureVideo` `ActivityResultContract` returns a
booleans to your `ActivityResultCallback` indicating if your video was
successfully saved to the given URI."
Test: added case to activity integration app
Bug: 185938070

Change-Id: Ie21f23fd680c361242c7ec11aa0684b9a6cdbf1a