Fix popUpTo+restoreState when inclusive=false
In cases where you use popUpTo and inclusive=false,
the back stack state needs to be associated not only
with the destination you've directly passed to
popUpTo, but also with the destination that was
actually popped.

In addition, this state also needs to be associated
with the parents of those destinations specifically
when the destination is the start destination of the
graph so that navigating to the graph itself is
treated the same as navigating to the start destination
when it comes to restoring state.

E.g., let's say you have three nested graphs: A, B, and C,
each with a respective start destination of A', B', and
C'. As per the Principles of Navigation, you should
always keep the start destination of your graph (A') on
the back stack. Therefore, the much more common case
is to use popUpTo(A') with inclusive=false when navigating
to B or C (or back to A). By adjusting the logic around
saving the state, navigating from B to C will save the
state of B to B, even though the popUpTo was set to A'.
Similarly, the state of A' (and notably, any destinations
stacked on top of it) will be restored when you navigate
to A with restoreState=true.

With this change, the BottomBarNavDemo for Navigation Compose
was updated to use the new restoreState and saveState flags
and now the state of the Scrollable destination is properly
saved and restored when you swap between the Scrollable
and Dashboard tabs by tapping on the bottom nav.

BUG: 80029773
Test: new NavControllerTest tests passes
Change-Id: I4ce337d269fe0db5ee424d5f47e64868c5a1a3a9