Refactor WatchFace construction to support fast metadata queries
Test: Manual testing
Relnote: We have introduced WatchFaceControlClient.getDefaultProviderPoliciesAndType which returns a map of Complication Ids to DefaultComplicationProviderPolicies and the default ComplicationType. Where possible a fast path is used that avoids fully constructing a watch face. To facilitate this the WatchFaceService API has had to change with two new methods being: createUserStyleSchema and createComplicationsManager the results of which are passed into createWatchFace. In addition Complications are now constructed with a CanvasComplicationFactory which allows for lazy construction of the CanvasComplication renderers.
Bug: 183188034
Bug: 184972867
Change-Id: Iad6c1414c4ff1792abaf2c60334fed74bd5b5686