Expose RotationReceiver as a public API
We would like to expose building blocks of CameraController. This is to bridge the feature disparity gap between camera-view and camera-core so that it's easier for developers to use PreviewView without CameraController. This will help MLKit team to build their MLKit-CameraX helper class.

- renamed the class to RotationReceiver because only inteface can be named .*Listener.
- updated java doc

Relnote: add a RotationReceiver class that receives Surface rotation changes. This can be used to set the target rotation when the device is in fixed portrait/landscape mode.

Bug: 188543457
Test: manual test and ./gradlew bOS
Change-Id: Ib278f5f6050370e6de8a492a22dd39c861155688