Ensure adapter position APIs work in ConcatAdapter for recycled VHs
This CL fixes a bug in ConcatAdapter where it was changing its internal
structure before calling onViewRecycled / onFailedToRecycle APIs. This
breaks documented API in onViewRecycled / onFailedToRecycle APIs where
it should be able to retrieve adapter position if the item is not
removed from the adapter.

I've moved callbacks before changing internal structure to fix the
issue.

Bug: 187339376
Test: ConcatAdapterTest
Change-Id: If5c9c182abed6d85df3b35e86ef79beed497c901