Create WatchFaces earlier in the flow
This was done inside OnSurfaceChange becuase we had to wait until
then for it to be possible to create the UI thread GL context for
GL watch faces. Now construction of the UI thread GL context is
deferred until well after the WallpaperService has created the
engine by which time onSurfaceChange has been called.

This means we don't have to wait before calling maybeCreateWCSApi.

WatchFaceControlClientTest has been modified to properly ensure
getOrCreateInteractiveWatchFaceClient is called before
onCreateEngine. Failure to do so lead to timeouts. Frustratingly
withTimeout { ... } wasn't working reliably in the tests (no
exception was thrown) so we now use a custom variant that does
correctly throw an exception.

Test: Manual testing
Change-Id: Ib12f36d59924fc86f9c8db7deae09bccab9ff481