Add builder wrappers for requests and events.
This is the first CL in a chain which will allow for the building of a
Tiles client lib, so developers can interact with a tile provider
themselves (thus, needing to build TileRequest etc).

This of course clashes with the existing reader classes (RequestReaders
et al); the next CL will migrate code usages to these new builder
classes and remove the Readers classes.

Bug: 186543504
Relnote: Added builder classes for Requests and Events.
Test: No new tests; new builders aren't currently exercised.
Change-Id: Ib5cf4cf39e98656ac8650207953faaf756a8c1c7