Move LazyColumn/Row offsets calculation to the measuring phase
Before this change the items offsets were calculated during the layout phase. This makes sense, but a bit incorrect as we set the new LazyListState.layoutInfo which contains the list of items with offset getters during the measure pass. This means that the assigned layoutInfo is not yet fully calculated and is effectively mutable when the listeners could already use it. After the current refactoring the LazyListMeasureResult object is fully calculated when returned from measureLazyList. Plus it has less properties in it as we keep things like headers or notUsedButComposedItems as the implementation detail of the measuring logic.
Second point is about finding and composing the sticky header during the layout phase. It is incorrect as with this approach LazyListState.layoutInfo does not contain the header because it would be composed later. This is fixed by this change, plus we were able to convert LazyListHeader into the stateless function.

Test: existing tests in lazy package. changed header tests to assert the header item is a part of LazyListState.layoutInfo
Bug: 185916470
Relnote: n/a
Change-Id: I015575e5efbbfad155e1f6e65d7eb436e65a6b60