Remove the tap detector used to clear focus
This makes compose similar to the classic view system where
clicking on white space does not clear focus. It also solves
these immediate issues:

1. The tap detector in FocusManager was preventing Android views
behind a Compose View from getting any clicks.
2. We wouldn't clear focus when the user clicked on whitespace
if the container consumed the touch events.

The selection container does not intercept clicks, and instead
of dismissing the selection when the user clicks the selection
container, it relies on focus being cleared as a means to
clear selection. Since this CL removes this feature. It adds a
detector to selection container to clear selection. The detector
is added only when the selection popup is shown and is removed
when the selection popup is not visible.

Bug: 189160953
Bug: 183631681
Bug: 179602539
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.ClearFocusTest
Test: ./gradlew compose:foundation:foundation:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.foundation.text.selection.SelectionContainerTest
Change-Id: Ic2636589623a048b1d4431c9727d4fb2922fa3e4