Remove ManualFrameClock from public API
The purpose of ManualFrameClock was to be able to control animations by
controlling the clock on which they run. We now have a better tool,
composeTestRule.mainClock, and ManualFrameClock is almost entirely
unused. It is therefore better to remove this from public API.

Bug: 190198467
Bug: 189951065
Test: ./gradlew bOS

Relnote: "Removed `ManualFrameClock`. If you need to control animations,
use `composeTestRule.mainClock` instead."

Change-Id: I3c3e8d0387c37ab3f3a29b648429056ac0eb6b26