Fixes ripples sometimes being instantly cancelled
If RippleForeground exits on the exact same millisecond as it entered (according to AnimationUtils.currentAnimationTimeMillis), it will instantly fade out instead of calculating the minimum amount of time it should show for. As we 'buffer' Interactions in clickable since we are not sure if a down event will turn into a scroll, a fast tap can (rarely and randomly) cause both PressInteraction.Press and PressInteraction.Release to be emitted on the same millisecond, causing this behavior.

Theoretically we can consider delaying the PressInteraction.Release (this is similar to what View.java does in the case of a 'prepress'), but given that some usecases might want to know that it was immediately emitted and not show any indication / state change in such a case, for now it seems better to just add a delay in RippleHostView if needed.

Fixes: b/190076611
Test: Manual
Change-Id: I365ff79e4985aff2506d823fef69ad9f72f0e264