Ensure that NavGraphNavigator stack order remains correct
When popping the root graph, the root graph
needs to be added to the front of the
hierarchy to ensure that the
navigatorBackStack for the NavGraphNavigator
is rebuilt in the correct order.

This ensures that future popBackStack() calls
will find that the NavController's back stack
ordering matches the NavigatorState of the
NavGraphNavigator, avoiding an
IllegalStateException.

Test: new NavControllerTest test passes
BUG: 190265699
Relnote: "Fixed an IllegalStateException when
calling popBackStack() after previously popping
the root graph of the NavController."

Change-Id: I2a330acdd339aece36455aa0da37d0609767552c