Fix @null support for StringArrayType
All NavTypes should support @null values and
StringArrayType is no exception. Ensure that
the Kotlin version of StringArrayType is
properly annotated to indicate that all methods
of that type support null values.

Relnote: "Fixed a null when navigating to a destination
using a `string[]` argument with a default value of `@null`."
Test: updated test suites pass
BUG: 190893500

Change-Id: I1fbe8da49d9652a842852778e9935e2426984ea9