Do not crash in a rare case when alignments are queried
It was happening when the grandparent is querying the alignments and the child is added lazily and we do not measure this child at all. This child is still scheduled for remeasure, but in most cases it will be skipped as this node is not placed. But not when the parent was asked for the alignments.
The fix is to just essentially do a no-op instead of a crash when we try to remeasure the node which was never measured before. It is safe to do so as this node can't affect any other node anyway unless it will be measured and placed next time.

Test: new test in AlignmentLineTest
Fixes: 189965769
Change-Id: I016a730ab0b60a41e264254e2f2ddb77aeb48d28