Add tests for imageBackgroundChip content colors.
Also restructure and rename some of the other Chip tests to allow for checking "content" colors only when there is a gradiant background.
Will follow up with Golden image tests for the Gradiant and ImageBackground chips.
Renamed some of the existing tests to make the names clearer.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "Added more Chip tests to cover the content colors for imageBackgroundChips"
Bug: 188882753
Change-Id: Ia9183927d1b1bd37da6e2b6a76498e7cbe320cf5