Added layout checkers to Wear Tiles' renderer.
These checkers will validate common issues with Tiles, and log errors to
logcat when using the viewer. At the moment, this only includes a
checker to ensure that at least one element has an accessibility
description assocaited with it, but can be easily expanded to cover more
cases in future.

Bug: 186542358
Test: ./gradlew :wear:tiles:tiles-renderer:build
Relnote: Added layout checkers to Wear Tiles' renderer.
Change-Id: I3a8692e72398b01fdb29a19b7f64d6d9452032de