Fix ActivityResultRegistry save/restore
The ActivityResultRegistry is currently restored after the on
ComponentActivity call to super.onCreate(). This means that any
callbacks registered before that could have already been saved waiting
to be restored so we end up improperly registering several duplicate
callbacks.

We should handle this on two fronts, one in ActivityResultRegistry for
cases outside of ComponentActivity and inside of ComponentActivity.

For ComponentActivity we should  save earlier as part of an OnContextAvailableListener to ensure that we restore the callbacks as soon as possible.

For ActivityResultRegistry, we should make sure that even if we register
before a call on onRestoreInstanceState, we don't duplicate keys.
Specifically, we need to restore the old requestCode to ensure that any
launched results that could potentially be dispatched still have a valid
requestCode, otherwise the dispatch would be ignored.

RelNote: "The ActivityResultRegistry callbacks are now properly saved
and restored so callbacks are not duplicated in the savedState."
Test: Added new test
Bug: 191893160

Change-Id: I9781617370ad24f768249df42d2ab148915097cb