Bring back the checkSuccess() method in AppSearchBatchResult.
Despite it being hidden, some of our users have started to use it. This
CL is the quickest way to unblock the sync. The alternative would be to
reproduce this method for each of our users, or to rely on
isSuccess().isTrue() which does not produce a very useful error message.

Change-Id: I6b0513c9430515fd29640ea2c4060518caa4e010