The ViewFactoryHolder should return itself as viewRoot
We used to report the uniqueDrawingId of the AndroidViewsHandler
with the ComposeView (child of AndroidView) when a legacy View was
embedded in a compose app. The id was used to position the child
AndroidViewsHandler under the AndroidView node in the layout
inspector.

However what I didn't realize is that the AndroidViewsHandler is
shared between all legacy Views under a single AndroidComposeView.
When there are multiple AndroidViews in a compose app we would move
all legacy Views under one of the AndroidViews in the layout
inspector.

Fix this problem be reporting the uniqueDrawingId of the
ViewFactoryHolder view instead.

Bug: 191695903
Test: Added unit test LayoutInspectorTreeTest
Change-Id: I260a627906ce41aadb79b7342f018ff2874437fa