getOrCreateInteractiveWatchFaceClient to always set the style
If after editing a watch face style, WCS crashes or is killed
before the edit is persisted then the watch face and WCS may be out
of sync.  When WCS restarts it calls
getOrCreateInteractiveWatchFaceClient however the userStyle
parameter is currently ignored if the instance ID already exists.

This patch makes getOrCreateInteractiveWatchFaceClient always set
the style, which in the above scenario would revert the style
change.

Bug: 192844401
Test: Manual testing
Change-Id: I90d6b719f2574214b485b2801d4011e13af743fe