Eliminate duplicate reports, inspect called member's experimental scope
Hacky, implementation-dependent workaround for lint visiting a single usage
site multiple times when there are multiples of the same annotation in scope.

Also traverses the called member's parent hierarchy to compute whether it's
actually part of an experimental group due to explicit opt-in or inclusion.

Relnote: """RequiresOptIn lint check has improved parity with the Kotlin
    compiler's implementation of opt-in, adding support for nesting @OptIn
    within the scope of a @RequireOptIn-annotated member."""
Bug: 193170595
Fixes: 193110413
Test: RequiresOptInDetectorTest
Change-Id: I7cfc77384d25f4d10a5470583ba292f0d5fe22c0