Supporting guava ImmutableMultimap in Room.
This CL focuses on supporting the multimaps of type ImmutableListMultimap and ImmutableSetMultimap. The user is oly allowed these two types as opposed to the generic `ImmutableMultimap` to align with the practices on the official documentation which states: "Avoid direct usage of ImmutableMultimap as a type (as with Multimap itself). Prefer subtypes such as ImmutableSetMultimap or ImmutableListMultimap, which have well-defined equals(java.lang.Object) semantics, thus avoiding a common source of bugs and confusion."

If the user attempts to use the generic type, a warning will be displayed accordingly.

Bug: 187490856
Test: MultimapQueryTest.java
Change-Id: I08f4e71f55a1cee76d6d2bc09c2e69bb31811c1c