Add hasDefaultValue method to XExecutableParameterElement
This CL adds a new method to check if a parameter has a default value
or not.

This information is only useful when generating kotlin code.
When generating java code, caller needs to check the @JvmOverloads
annotation before assuming that the method can be called from Java
sources without providing the parameter.

Note that how JvmOverloads are handled between KSP and KAPT is
different.
In KAPT, XProcessing will have 1 method for each overload but only the
original method's parameters might return `true` from hasDefaultValue
(we don't have a way to detect JvmOverloads, it does not show up neither
in stubs nor in metadata).
In KSP, we'll have only 1 method.

Bug: 169882122
Test: XExecutableElementTest
Change-Id: I07e47f87eb05891401369a6ca4d044da3fd2f42c