Add CurvedText component
A CurvedText component allows developers to easily write text following
the curvature a circle (usually at the edge of a circular screen).
Needs to be part of a CurvedRow

Mainly to discuss the API and "creative" uses of compose. In
particular, how to pass information between layout (CurvedRow) and child
components (CurvedText):
* The layout needs to tell the radial positioning (distance to the
center) to the child. It does so using the max dimensions passed to Measurable.measure()
* The layout needs to know the dimensions of the child to compute the
previous, before calling measure (since it can't be called twice). It uses the intrinsic
measures and an alignment line.

Note that this component only works on Android, for now.

Relnote: "Add CurvedText component, that allows developers to easily write text following
the curvature a circle (usually at the edge of a circular screen)"
Test: Added a screen to the Demo, unit + screensot tests comming in later CLs
Bug: 190806710

Change-Id: Id1267f2b4bc88780b7a91b77a51c31d3606fc9aa